chat-coordinator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
aiohttp==3.9.1
pydantic==2.5.0
asyncio-throttle==1.0.2
